Removed mixins
