+++ TOMsPlugin/ui/proposalPanelDockwidget.py
@@ -26,7 +26,7 @@ class ProposalPanelDockWidget(QDockWidget, FORM_CLASS):
 
     def __init__(self, parent=None):
         """Constructor."""
-        super(ProposalPanelDockWidget, self).__init__(parent)
+        super().__init__(parent)
         # Set up the user interface from Designer.
         # After setupUI you can access any designer object by doing
         # self.<objectname>, and you can use autoconnect slots - see
